Remove contact form
Netlify is having some difficulty with the App router. I found
[this thread](https://answers.netlify.com/t/server-edge-not-defined-error-on-nextjs-ssr-functions-cause-site-to-return-500-errors/91793/116),
which seems to indicate that the problem is on Netlify's end. But this
is just a placeholder site, so it's not worth losing to much time on.

The next step will likely be to build a site with 11ty, which is a better
fit for this project. But if I do decide to keep using Next, I'll likely
switch the host to Vercel, which seems better suited to Next.